requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
python-multipart==0.0.12

# Config
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Redis
redis[hiredis]==5.1.0

# AI / Translation
groq==0.11.0

# STT
openai-whisper==20240930

# TTS (fallback)
edge-tts==6.1.12

# QR code
qrcode[pil]==8.0

# Retry logic
tenacity==9.0.0

# Logging
structlog==24.4.0

# Tests
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
